+++ src/LDL/Validators/IntegerValidator.php
@@ -2,6 +2,7 @@
 
 namespace LDL\Validators;
 
+use LDL\Framework\Base\Contracts\Type\ToIntInterface;
 use LDL\Validators\Exception\TypeMismatchException;
 use LDL\Validators\Traits\NegatedValidatorTrait;
 use LDL\Validators\Traits\ValidatorDescriptionTrait;
@@ -29,6 +30,10 @@ class IntegerValidator implements ValidatorInterface, NegatedValidatorInterface
 
     public function assertTrue($value): void
     {
+        if($value instanceof ToIntInterface){
+            $value = $value->toInt();
+        }
+
         $valid = is_int($value);
 
         if($valid && !$this->unsigned){
@@ -44,8 +49,9 @@ class IntegerValidator implements ValidatorInterface, NegatedValidatorInterface
         }
 
         $msg = sprintf(
-            'Value expected for "%s", must be of type integer, "%s" was given',
+            'Value expected for "%s", must be of type integer or an instance of "%s", "%s" was given',
             __CLASS__,
+            ToIntInterface::class,
             gettype($value)
         );
 
@@ -54,13 +60,14 @@ class IntegerValidator implements ValidatorInterface, NegatedValidatorInterface
 
     public function assertFalse($value): void
     {
-        if(!is_int($value)){
+        if(!is_int($value) && !$value instanceof ToIntInterface){
             return;
         }
 
         $msg = sprintf(
-            'Value expected for "%s", must NOT be of type integer, "%s" was given',
+            'Value expected for "%s", must NOT be of type integer or an instance of "%s", "%s" was given',
             __CLASS__,
+            ToIntInterface::class,
             gettype($value)
         );