+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1200203297621230 - Add HasValidatorResultInterface
 - feature/1200836379007029 - Add StringLengthValidator / StringContainsValidator
 - feature/1201055143148093 - Add example outputs. Delete run.php
+- feature/1201235392821202 - Add type interfaces support for validators
  
 ### Changed